+++ lab3/command.h
@@ -1,0 +1,4 @@
+void reboot();
+void loadimg();
+unsigned int get_system_frequency();
+unsigned int get_system_count();
