+++ lab3/string.h
@@ -1,0 +1,3 @@
+void itoa(long result);
+void clearbuf(char *buf, int size);
+int strcpy(char *buf1, char *buf2);
